Implemented permission functionality
When the profile is accessed through the landing page, the profile will display all of the user's info to them. When the profile is accessed through the employee list frame, it will show the selected employee's info. If the user is an admin, they will see all of the selected employee's info. If the user is another employee, they will see only some of the selected employee's info